Remove icons
Taken from https://github.com/alloy/terminal-notifier

"The Notification Center always uses the application’s own icon, there’s currently no way to specify a custom icon for a notification. The only way to use this tool with your own icon is to include a build of terminal-notifier with your icon instead."

Since we use terminal-notifier we can't use icons for this script. Simple as that. Once Eloy or anyone on GitHub found a way to do so I will update the script and let the user define an icon.

Signed-off-by: Patrick Kontschak <patrick.kontschak@gmail.com>