Remove activate option
It would be annoying for the user if Terminal.app would launch whenever they click on the notification. If people are getting annoyed that nothing happens whenever they click on the notifications, feel free to create a pull-request with an check what Terminal app the users uses (not everyone likes Apple's Terminal.app, some people prefer iTerm2, etc.).

Signed-off-by: Patrick Kontschak <patrick.kontschak@gmail.com>